refactor mpr toggle and stack sync
