+++ extensions/cornerstone/src/utils/mpr/getProtocolViewportStructureFromGridViewports.ts
