+++ extensions/cornerstone/src/utils/stackSync/calculateViewportRegistrations.ts
