+++ platform/viewer/public/config/default.js
@@ -4,7 +4,7 @@ window.config = {
   extensions: [],
   modes: [],
   showStudyList: true,
-  // maxNumberOfWebWorkers: 3,
+  maxNumberOfWebWorkers: 4,
   // below flag is for performance reasons, but it might not work for all servers
   omitQuotationForMultipartRequest: true,
   showLoadingIndicator: true,
